Added navigation file without seperate namespace
