+++ src/public/index.ts
@@ -21,9 +21,6 @@ export {
   getTabInstances,
   initialize,
   initializeWithFrameContext,
-  navigateBack,
-  navigateCrossDomain,
-  navigateToTab,
   print,
   registerBackButtonHandler,
   registerBeforeUnloadHandler,
@@ -37,9 +34,9 @@ export {
   setFrameContext,
   shareDeepLink,
 } from './publicAPIs';
+export { returnFocus, navigateBack, navigateCrossDomain, navigateToTab } from './navigation';
 export { settings } from './settings';
 export { tasks } from './tasks';
-export { navigation } from './navigation';
 export { ChildAppWindow, IAppWindow, ParentAppWindow } from './appWindow';
 export { media } from './media';
 export { location } from './location';
