+++ src/public/publicAPIs.ts
@@ -12,7 +12,6 @@ import { settings } from './settings';
 import {
   TabInformation,
   TabInstanceParameters,
-  TabInstance,
   DeepLinkParameters,
   Context,
   LoadContext,
@@ -268,18 +267,6 @@ export function registerBackButtonHandler(handler: () => boolean): void {
 }
 
 /**
- * Navigates back in the Teams client. See registerBackButtonHandler for more information on when
- * it's appropriate to use this method.
- */
-export function navigateBack(onComplete?: (status: boolean, reason?: string) => void): void {
-  ensureInitialized();
-
-  const messageId = sendMessageRequestToParent('navigateBack', []);
-  const errorMessage = 'Back navigation is not supported in the current client or context.';
-  GlobalVars.callbacks[messageId] = onComplete ? onComplete : getGenericOnCompleteHandler(errorMessage);
-}
-
-/**
  * @private
  * Registers a handler to be called when the page has been requested to load.
  * @param handler The handler to invoke when the page is loaded.
@@ -313,29 +300,6 @@ export function registerChangeSettingsHandler(handler: () => void): void {
 
   GlobalVars.changeSettingsHandler = handler;
   handler && sendMessageRequestToParent('registerHandler', ['changeSettings']);
-}
-
-/**
- * Navigates the frame to a new cross-domain URL. The domain of this URL must match at least one of the
- * valid domains specified in the validDomains block of the manifest; otherwise, an exception will be
- * thrown. This function needs to be used only when navigating the frame to a URL in a different domain
- * than the current one in a way that keeps the app informed of the change and allows the SDK to
- * continue working.
- * @param url The URL to navigate the frame to.
- */
-export function navigateCrossDomain(url: string, onComplete?: (status: boolean, reason?: string) => void): void {
-  ensureInitialized(
-    FrameContexts.content,
-    FrameContexts.sidePanel,
-    FrameContexts.settings,
-    FrameContexts.remove,
-    FrameContexts.task,
-  );
-
-  const messageId = sendMessageRequestToParent('navigateCrossDomain', [url]);
-  const errorMessage =
-    'Cross-origin navigation is only supported for URLs matching the pattern registered in the manifest.';
-  GlobalVars.callbacks[messageId] = onComplete ? onComplete : getGenericOnCompleteHandler(errorMessage);
 }
 
 /**
@@ -391,19 +355,6 @@ export function executeDeepLink(deepLink: string, onComplete?: (status: boolean,
   ensureInitialized(FrameContexts.content, FrameContexts.sidePanel, FrameContexts.settings, FrameContexts.task);
   const messageId = sendMessageRequestToParent('executeDeepLink', [deepLink]);
   GlobalVars.callbacks[messageId] = onComplete ? onComplete : getGenericOnCompleteHandler();
-}
-
-/**
- * Navigates the Microsoft Teams app to the specified tab instance.
- * @param tabInstance The tab instance to navigate to.
- */
-export function navigateToTab(tabInstance: TabInstance, onComplete?: (status: boolean, reason?: string) => void): void {
-  ensureInitialized();
-
-  const messageId = sendMessageRequestToParent('navigateToTab', [tabInstance]);
-
-  const errorMessage = 'Invalid internalTabInstanceId and/or channelId were/was provided';
-  GlobalVars.callbacks[messageId] = onComplete ? onComplete : getGenericOnCompleteHandler(errorMessage);
 }
 
 export function setFrameContext(frameContext: FrameContext): void {
